Converting game objects to python dictionaries
Added conversion of the game objects to python dictionaries. This will allow for a much easier time finding properties of objects, without having to craft ridiculous regex. This in particular, is aimed towards the automation of parsing the "effects", allowing for automated effects tables for units. The code is not perfect, and it breaks on some edge cases, but it should work for most units.